backend/requirements.txt
asgiref==3.5.0
backports.zoneinfo==0.2.1
Django==4.0.4
django-cors-headers==3.11.0
djangorestframework==3.13.1
djongo==1.3.6
dnspython==2.2.1
pykerberos==1.2.4
pymongo==3.12.3
python-snappy==0.6.1
pytz==2022.1
sqlparse==0.2.4
